docs (small refactoring)
